KUDU-776: Improve handling of unresolvable master hostnames
Before this patch, passing an unresolvable master address to the java client ended up in a whole
lot of log spew. The reason is that we weren't trying to getIP before creating a TabletClient
for a master and thus used the passed hostname to connect(). Then, getting an error, we would try
to find the TabletClient's hostname that gave us the error and fail again.

This patch makes it so that we try to resolve the passed hostname before we try connecting to it
so that we error out right away.

Change-Id: I21c348a8a796ca9e157b4a2073b8bfc3b15e5b92
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6670
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins