KUDU-764: Abandon RowsWithErrorException and move the RowErrors to OperationResponse
It's currently not very obvious how to get to the row errors, and as a first step towards
KUDU-764's goal of adding server-provided errors to the RPC responses, this patch moves the list
of RowErrors from RowsWithErrorException (which gets discarded) to OperationResponse.

The advantage of doing this is pretty obvious from the patch. The handling of KUDU-568 is
pushed down into the sessions, disabled by default.

Change-Id: I0556179ca627a6e31090b0c10cb5fe046447862a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6676
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins