subprocess: allow multiple calls to Wait()
It's convenient to use WaitNoBlock() to see whether a subprocess has crashed,
for example for asserting that minicluster daemons haven't crashed during some
integration test. This patch changes Subprocess so that repeated calls to
Wait() or WaitNoBlock() simply re-return the previous exit code if the subprocess
has already been reaped.

Change-Id: I2e743daca0c640a976ac33066538b3c5326362a0
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6677
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins