KUDU-775. Leader should not crash if follower falls behind GC point
This fixes a FATAL if the follower has been down long enough that the
leader has GCed the logs necessary to catch them back up.

For now, we still WARN on this case. Eventually we'd like to do something
smarter about having the master remove this replica and recruit a new one.

There is one TODO mentioned in the test: I think even with this fix, the
abandoned follower will keep disturbing the other nodes trying to call
elections. It seems likely that the fix for that is the election pre-check,
and we can assume that once we have 'RemoveServer', we'll deal with the
abandoned follower a bit more harshly.

Change-Id: I4ce5b7275ffec9339718c94ba6e587c08bbb22a6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6678
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Reviewed-by: David Alves <david.alves@cloudera.com>