KUDU-562. Prevent a partitioned/abandoned node from disturbing an active leader
This implements section 4.2.3 of Diego Ongaro's Raft thesis[1] which suggests
that servers should deny votes when they have recently heard from an active
leader (or are themselves an active leader).

The implementation approach taken is a simple one based on the implementation
in LogCabin[2]. RaftConsensus maintains a timestamp before which it won't
ever vote in an election. The timestamp is bumped forward by the minimum
election timeout when it hears from a leader. It's set to a maximum value
(higher than all other times) when the local node is an active leader. It's
set to a minimum value (below all other times) when first becoming a replica.

As suggested in the thesis, the RequestVote() RPC has a new boolean flag which
can skip this check. This allows for faster leadership transfer when we explicitly
want to move leadership from one node to another. It also comes in handy in
several unit tests where we expect that we can trigger a leader change on a
quiesced quorum by asking one of the followers to start a vote.

Note that there is one downside to this approach vs a "pre-election" round-trip
in which the candidate checks preconditions before incrementing its term:
if a node is partitioned from the others for some amount of time, it will
increment its term several times. Then, when connectivity is restored, the
leader will see the higher term and step down, causing an unnecessary election.
This seems like a rare enough scenario that the simpler approach is OK for now;
we'll observe how this behaves on real clusters and implement the pre-election
round trip at a later date if necessary.

The modified unit test verifies that a node that has been abandoned by the
active leader can't disturb the healthy leader. The test verifies this by
looking at the leader's term. I also manually verified that the logs contain
the new status message in response to the election attempts.

[1] https://ramcloud.stanford.edu/~ongaro/thesis.pdf
[2] https://github.com/logcabin/logcabin/blob/master/Server/RaftConsensus.cc

Change-Id: Ia4c1d15480d3cff08626ae15e777e1ab441eae3b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6680
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Reviewed-by: David Alves <david.alves@cloudera.com>