consensus: Replace last_sent in the queue with per-term last received
There is a problem with the current queue implementation where if a
follower temporarily refuses an op, and needs it to be re-replicated,
the leader will not resend it if there is a heartbeat in between the
rejection and the next append. This is mainly due to the fact that
last_sent seems to be broken substitute for a properly working
last_received.

This change removes last_sent, and adds an additional optional parameter
to ConsensusResponsePB called last_received_current_leader. This field
is reset to MinimumOpId() every time there is a term change.

In addition, this change replaces log_tail in TrackedPeer with
next_index and significantly simplifies the logic in
PeerMessageQueue::RequestForPeer() while moving all of the logic around
how to set next_index and last_received into
PeerMessageQueue::ResponseFromPeer().

On the leader queue side, the current last_received for each peer (which
controls both the watermarks and the next op to send) is chosen based on
the following logic: if last_received (the latest OpId in the
follower's log) exists in the leader's log, that is considered to be the
"real" last_received for that peer. If not, we consider the peer log to
have diverged, and we fall back to the value of the
last_received_current_leader OpId.

Existing tests were modified to continue functioning using this logic,
and a new test was added to verify the per-term resetting behavior in
raft_consensus-test.

An alternative approach would be to attempt to actively truncate the log
on term change, but that may require additional log APIs and also result
in a common scenario of duplicates on every term change. So I went with
this approach but an open to alternatives.

I considered adding a test for the specific behavior of the leader queue
when the follower partially fails to send, but the test seems tricky to
set up and this behavior is implicitly exercised by the AddServer /
RemoveServer tests so it seems reasonable to rely on those as they get
merged.

As part of this change, made LogCache::LookupOpId() public so that the
queue can check whether a given last_received OpId is in its log.

Also changed ASSERT_OPID_EQ slightly so that it's possible to append
more messages to the assertion if it fires.

Change-Id: I22943fdd9063014054eca57683568a9d05a927d0
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6665
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: David Alves <david.alves@cloudera.com>