Don't look at empty DMS when trying to find the best one to flush
We saw a weird behavior on a cluster with long flushes (30+ seconds), DMS were being
flushed non-stop even though they were empty when their tablet was being MRS-flushed.

We found that Tablet::FindBestDMSToFlush, reading the list of rowsets, was getting the MRS
since it gets inserted into that list while its flush is happening. Its DeltaMemStoreSize
is 0 but it does have a log retention size which can be pretty high, and in our case it was
always the highest.

The fix is as simple as skipping empty DMS (the MRS always returns true when asked), but
testing it is harder since it requires a half-flushed state. It was tested with the
multi-threaded tpch insert.

Change-Id: I6bf48d214c5a5b2782b298381e836b48fa423b14
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6689
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins