raft_consensus: initialize withhold_votes_until to Min()
In 4870ef20b9a0c729c1d7f16c016c7d91b193e46f I added this member variable
but did not initialize it in the constructor. This meant that, if a
replica had started and not received any UpdateConsensus(), but then received
a vote request, it would fail with a CHECK failure like:

F0520 14:17:13.575253 48415 monotime.cc:195] Check failed: rhs.Initialized()
*** Check failure stack trace: ***
    @     0x7f61a4d36bca  google::LogMessage::Fail() at ??:0
    @     0x7f61a4d36496  google::LogMessage::SendToLog() at ??:0
    @     0x7f61a4d368ee  google::LogMessage::Flush() at ??:0
    @     0x7f61a4d3a08c  google::LogMessageFatal::~LogMessageFata
l() at ??:0
    @     0x7f619f0a9fe3  kudu::MonoTime::ComesBefore() at ??:0
    @     0x7f61a2d25c39  kudu::consensus::RaftConsensus::RequestVote() at ??:0
    @     0x7f61a4fc3895  kudu::tserver::ConsensusServiceImpl::RequestConsensusVote() at ??:0
    @     0x7f61a26593d1  kudu::consensus::ConsensusServiceIf::Handle() at ??:0
    @     0x7f61a1fcfdb3  kudu::rpc::ServicePool::RunThread() at ??:0
    @     0x7f61a1fd0bae  boost::detail::function::void_function_obj_invoker0<>::invoke() at ??:0
    @     0x7f619f0cc5ce  kudu::Thread::SuperviseThread() at ??:0
    @       0x3346407851  (unknown) at ??:0
    @       0x33460e894d  (unknown) at ??:0

This is causing a lot of flakiness on raft_consensus-itest.

This solution is to initialize it to 'MonoTime::Min', such that, if a replica
has never heard from any leader, it will happily vote in an election.

Change-Id: Ib59436532a734c5514a65def071a79b37411e424
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6693
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>