CSD work to enable basic Kudu monitoring by CM.
This change hooks calling generate_mdl.py into the CSD build and includes the
output in the Kudu CSD. The script has been tweaked to allow specification of
an output file since this was easier for me to do than get piping to work in
maven. I also made some adjustments to the context of the generated metrics
that the CM agent needs. Finally, I've added the flags for the web UI listen
interfaces and emitted these and the port to a dedicated file used by the CM
agent to discover where it ought to be polling for metrics.

Testing done: I built the CSD, put it on my cluster (thanks Adar!) and I see
metric colleciton working end-to-end.

Change-Id: Iff995f1b11f801ab96aa232f4d50d0efe4b5a508
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6691
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>