Move the libcurl dependency to the kudu_test_util library
The exported client was unecessarily importing libcurl, which in turn
meant that it could import a different version of libldap than impala's,
which in turn meant that this bug would occur, in some machines:

https://bugzilla.redhat.com/show_bug.cgi?format=multiple&id=755725

Since libcurl is only used in tests this moves it to the kudu_test_util
library. I've also added this library as a dependency to the twitter
demo, which also uses curl.

Change-Id: I7a0908bbfb1b39f644aed717da89e3921272a0a7
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6690
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins