KUDU-764: Sending the row errors to each callback individually
AsyncKuduSession had a TODO about not sending the whole OperationResponse that came from
a batch to all the individual Operation callbacks. Binglin also raised the same concern to
us recently.

This patch breaks out OperationResponse into two classes: OperationResponse which is the rpc
response for a single Operation and BatchResponse which is the rpc response for a single Batch.

Change-Id: I85a8d04be6656fa39285b12874d30abf6369a445
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6694
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>