raft_consensus-itest: test for disturbing node should only pause a follower
This fixes a major source of flakiness for the newly-added
TestFollowerFallsBehindLeaderGC test. The test was pausing a random tablet
server after starting. This could have been the elected leader, in which case
our references to 'leader' later in the test were no longer correct.

The new implementation still picks an arbitrary node to pause, but then
it waits until it sees a leader that does not match the node it picked.
This ensures that the node it thinks is 'leader' is in fact the leader
while performing writes, etc.

Right before resuming the paused node, we now make a note of the term.
This term may not be '1' in the case that our pausing caused a leader
election at the start of the test. We then resume the paused replica,
wait a few seconds, and verify that the term doesn't increase from what
we noted.

Before this fix, the test was flaky 30%+ of the time. With the fix,
it failed 30/1000 runs. The next commit in this series improves that.

Change-Id: If7aaebed7911a535e1140930aa762620258135cb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6695
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>