raft_consensus-itest: fix waiting for log GC
The code that tried to wait for log GC was listing the children
of the WAL dir, but the actual WAL segments are inside a tablet
directory _within_ that WAL dir. This fixes the waiting code to
look in the correct place.

This fixes the remaining source of flakiness in the
TestFollowerFallsBehindLeaderGC test case. I looped it 1000 times
and it passed 100%.

Change-Id: I8db2aaa97f8d93b0c620a309533bb3079b8634ad
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6697
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: David Alves <david.alves@cloudera.com>