raft_consensus-itest: disable async segment preallocation
This fixes one more super-rare (1/1000) case of flakiness I saw on this
test when running on a loaded machine:

In the failure case, the 13MB of data that we write only generated
two log segments because async preallocation happened too slowly.
Because we only had two segments, we were not able to GC the first
segment, and the node that was supposed to be "kicked out" of the
cluster was able to continue running.

Disabling preallocation appears to have fixed this flakiness.

Change-Id: Ib6f3d073e619718481170c0178e1f5b2d0ee0b34
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6707
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>