KUDU-783. Fix failure to bootstrap after tumultuous leader change
This fixes an issue spotted occasionally in raft_consensus-itest where
one node fails to bootstrap its tablet because it sees an out-of-order
operation.

The sequence of events that causes this is:

1) A replica has ops 1.1 through 1.10. Only 1.1 is committed.

2) A new leader, which only has up to op 1.5, is elected. It appends its NO_OP as
2.6 to its log, and sends the initial status-only requests to the peers. The
status-only request has a preceding opid of 2.6. (not sure if this is a change
since the NO_OP stuff or it was always this way). The request in this case has
no operations.

3) The replica hits this code path in the LMP function:
    // If the terms mismatch we abort down to the index before the leader's preceding,
    // since we know that is the last opid that has a chance of not being overwritten.
    // Aborting preemptively here avoids us reporting a last received index that is
    // possibly higher than the leader's causing an avoidable cache miss on the leader's
    // queue.
    if (term_mismatch) {
      return state_->AbortOpsAfterUnlocked(req.preceding_opid->index() - 1);
    }
this causes it to abort 1.6 through 1.10, but not append anything else (since
it was a status-only request).

4) The term-2 leader then crashes.

5) Some other leader, which did have ops 1.6 through 1.10, gets elected for
term 3.  It then manages to replicate those to the replica.  The replica is
left with 1.1 through 1.10, followed by a "repeat" of 1.6 through 1.10 in its
log. This makes it fail to bootstrap with an "Unexpected" op id error.

The fix here is quite simple: we just change the check during bootstrap to
be more permissive.

The testing is a bit more complex. I was unable to make any deterministic test
reproduce this issue. However, I added a very simple fault injection flag
which makes the leader crash with some probability right before sending
consensus RPCs. The test then runs this faulty cluster through a set number
of crash/elect/restart scenarios, verifying that replicas converge at the
end. This test failed about 5% of the time due to the above issue. With
the fix, I looped it 1000 times and it passed 1000/1000.

It seems that the new test interacts poorly when run in conjunction with some
other heavy stress tests in the integration-tests/ directory. I marked several
of the more heavy-weight multi-threaded tests as RUN_SERIAL true to try to
avoid these interactions.

Change-Id: I47ec302f8bcce72a60de8a253b616a05527245bb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6718
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>