external_mini_cluster: Bind TS HTTP server to pid-specific local IP
The tablet server already uses a pid-specific localhost IP (127.x.y.z)
for RPC, but wasn't using that IP for its web server. When I was looping
some consensus itests thousands of times, I'd occasionally hit failures
to bind the web server.

After this patch, I no longer see the issue.

We should separately do this for the master, but the master wasn't
blocking me, so not doing it in the same patch.

Change-Id: Id82a80a0b159afe249a50c1cf0be6ae6afb27246
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6717
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>