mm: better caching for tablet ops stat collection
Turns out that collecting stats for maintenance manager ops can be quite
expensive on a large deployment. For example, we've observed
MaintenanceManager::FindBestOp() consume >5s on ~5 TB of data, which exceeds
the naive time-based caching found in CompactRowSetsOp.

So let's do something smarter: let's cache the stats forever and invalidate
them when appropriate using tablet metrics. I've added caching to all ops
except FlushMRSOp, LogGCOp, and FlushDeltaMemStoresOp. The first two have
cheap implementations of UpdateStats() while the last is more expensive but
also more difficult to cache (we'd need to invalidate when any DeltaTracker
arena grows in size).

The caching reduces the steady-state overhead of FindBestOp() from 100% of a
core to about 25% on a2414.

Change-Id: I31ceb026c762c90bd4f646efede93a8ae585d229
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6730
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>