delta_tracker: convert component_lock into spinlock
This lock is never held for very long, so let's make it a spinlock instead
of a sleeping mutex. Doing so reduces the steady-state overhead of
MaintenanceManager::FindBestOp() from 25% of a core to 10% using a 5TB
tserver running on a2414.

While I was there, I converted compact_flush_lock into a Kudu mutex.

Change-Id: I989848b651be5a1af68d12a2d91b8f5a1f19a0bd
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6731
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins