consensus: Add RPC for simple leader stepdown
This adds a simple implementation of stepping down as leader in response
to an RPC.

This also changes BecomeReplicaUnlocked() to clear the current leader
flag.

This does not implement any of the potential optimizations for leader
stepdown, such as abdicating to a follower. We may want to support that
later to minimize unavailability, but this is enough to abdicate for
correctness-testing testing purposes.

Also add a bunch of helper functions in cluster_test_util, some moved
from existing itest code.

Also add a simple unit test for the above.

Change-Id: Ifa1e02fddfa6d3673b7cf078f609d6a02ed67b3e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6774
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: Michael Percy <mpercy@cloudera.com>