fs: allow FsManagers to be opened in read-only mode
When read-only, block creation/deletion is prohibited, as is new filesystem
creation. Useful for the FsTool family of tools.

I thought about making the read-only check return a bad status, but settled
on the CHECK because I think it's indicative of a programming error, and
because DCHECK may be insufficient (i.e. we have no unit test coverage for
many tools).

Change-Id: Id67e101364ae523f65c173d3143a82c2efcb1ec3
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6742
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>