raft_consensus: fix triggering of fast leader election at startup
We attempted to trigger a fast leader election on new tablets by
checking for term 1. However, a new tablet actually starts with
term 0. This corrects the check. I tested by verifying in the logs
that the leader election is started immediately.

This also brings the test time for raft_consensus-itest in short test
mode from ~1m5s down to ~49s.

Change-Id: I6da97d18b1fae43793da86a04855d37dec95fda0
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6747
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>