Trigger heartbeats immediately when tablet state changes
This speeds up the heartbeating of consensus configuration changes
by triggering an immediate heartbeat when a tablet is marked dirty.

It's difficult to test this, as it's a timing improvement, but I
added some temporary logging and saw the effect. I also noted that
the test time for raft_consensus-itest went from about 49s to about
39s.

Change-Id: I42d7026a003bf62f9954e9a55c5e1c2b0ca680c4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6748
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>