raft_consensus: the leader should notify followers aggressively about commits
This resolves a long-standing TODO that the leader should notify its followers
immediately if the commit index changed. This also makes the queue set 'more_pending'
to true not only when there are more messages in the log for the peer but also when
the committed index is greater than the peer's. Finally this changes the meaning
of 'empty queue' so that if a peer gets a request with no operations from the queue
but the committed index has moved, the request is considered non empty and the
peer still sends the request even if the 'force_if_queue_empty' wasn't set.

This is missing a test, but a test for this is quite hard because, though we have
an rpc to get the log tail, we don't have one for the committed index, which would
be required to build a such a test.

Change-Id: I2c0d8fd1efa875b3215f17a4749ce8720ead9cdd
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8275
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>