Add an LSAN suppression for atexit functions
For some unknown reason, ec2 slaves just started reporting a leak
on tablet_peer-test:

Direct leak of 1040 byte(s) in 1 object(s) allocated from:
    #0 0xc528ab in calloc /data1/jenkins-workspace/kudu-gerrit/BUILD_TYPE/ASAN/label/kudu-gerrit-slaves/thirdparty/llvm-3.4.2.src/projects/compiler-rt/lib/asan/asan_malloc_linux.cc:90
    #1 0x7f973868afa3 in __new_exitfn (/lib64/libc.so.6+0x35fa3)

It's not clear what changed that made this pop up, but it seems like
a reasonable thing to just suppress.

Tested this commit by running the test on one of our slaves.

Change-Id: I9b274b64dffa789ab40d21cec20c2920fb3b58a1
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6750
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>