java: put test output in TEST_TMPDIR or /tmp/kudutest-$UID
Doing this is more consistent with the rest of our tests, and is actually
pretty important if your source tree is running on ecryptfs, which doesn't
support hole punching (used by the log block manager).

I've also tightened up the @AfterClass file cleanup, which will let us
remove the groovy postbuild snippet in the kudu-gerrit job.

Change-Id: Ibf9a69a1c49234a194cc905d7ffa980ef25d7624
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6758
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins