raft_consensus: do not start leader elections when already LEADER
This fixes a race we observed on the YCSB cluster. It appears we hit the
following sequence of events:

- a node is a follower for term 1
- its failure detector files, causing it to start an election
- we snooze the detector and send out the leader election RequestVote() RPCs
- we get the leader election callback indicating that we should become leader
-- we acquire the consensus lock
-- the failure detector fires again, but the callback is blocked trying to
   acquire the lock
-- we call BecomeLeader() and start the consensus queue, etc, and unlock
   consensus.
- the blocked failure detector callback now proceeds, and it calls StartElection().
- StartElection() was calling state_->IncrementTermUnlocked(), which changed the
  term, but did not stop the consensus queue.

This caused the consensus queue to continue running as leader, even though the
term had advanced. The other peers would then crash as they received requests
in from this node falsely claiming to be leader in a term in which it wasn't
elected.

This patch makes two fixes:

1) Remove the IncrementTermUnlocked() call which directly modified the term.
This call was not safe to use, since it would not step down to FOLLOWER state
in the new term. Instead, we go through the same code path as other places
which change term, thus ensuring that we step down on term change.

2) StartElection() now short-circuits itself in the case that we are already
the LEADER of the current term. This part of the fix is not strictly necessary
for correctness - it's OK to step down and start a new election if you are a
LEADER. However, it avoids a spurious re-election.

The included integration test sets all of the failure detection intervals to be
extremely low, causing a lot of churn. This reproduced the bug very quickly
before the fix.

Change-Id: I22b7077439badf341989fcfc3e407f55d7b0cc1a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6766
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>