consensus: Add support for Raft membership change
Add support for one-by-one membership change as specified in Raft. This
replaces the existing (broken) ChangeConfig() RPC call.

This new call takes parameters to either add a server, remove a server,
or change the role of a server in the consensus group. At this time,
role change is not yet implemented. It's not difficult to add it, but
this patch is big enough as-is.

Also rename ChangeConfigRequestPB to ChangeConfigRecordPB, and continue
to use that in the logs to signify a config change. This is distinct
from the RPC calls, since the RPC call indicates a delta from the
current consensus group / quorum configuration, while the
ChangeConfigRecordPB records the results of applying that delta.

Tests added:

  - A test that verifies election and write-capability assumptions as
    we remove nodes and then add them back.
  - A test that verifies that a "pending" quorum member can be elected
    leader, as required by Raft.
  - A test that verifies that we can remove and add nodes while under a
    write workload.

Also:

Vote for candidates even if they are not in your local quorum. This is
required by Raft. Also make raft_consensus_quorum-test pass with unknown
candidate votes.

Fix some comments in consensus.proto and improve validation in
raft_consensus.cc

Change-Id: I6cc010f78a19df3be8dd8f6b10711330e7122b77
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6269
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>