KUDU-744. Randomize the exponential backoff after leader elections
Previously, we were only relying on randomization of elections from the
randomized failure detector. But, this randomization is an absolute time,
rather than a fraction of the timeout. So, even though we exponentiallly backed
off from bad leader elections, if each election is taking longer than the
randomization amount of the failure detector, we can get into lockstep.

This adds a multiplicative randomization factor to the actual backoff
calculation. We calculate the additional sleep time due to backoff, and then
multiply that by a fraction between 0 and 1. This ensures that, even after many
rounds of backoff, the nodes trying to be elected will fall evenly distributed
between the initial timeout and the maximum timeout.

I tested this using RaftConsensusITest.TestChurnyElections. Without the fix, a
few percent of the runs would end up with the nodes proceeding in lockstep and
not complete. With the fix, I ran 1000 iterations with no timeouts.

I also ran all of raft_consensus-itest 200 times. All passed except for one
failure due to an unrelated known issue.

Change-Id: I633cb5a6f41554be90d3d9a2ec6d584b6bc420fd
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6767
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins