mm: move memory limit to memtracker
The maintenance manager had its own memory limit and heuristic for
determining a reasonable limit if none was provided, but this always seemed
odd to me in light of the (newer) memtracker framework, which ostensibly
limited memory consumption in the entire process.

So in this patch, the mm memory limit moves to the root memtracker. Further,
the consumption value of the root memtracker now drives the emergency flush
behavior in the mm.

The latter change deserves some thought: although all of the anchored RAM in
the ops is also present in the root memtracker, there's other stuff in the
root too. As a contrived example, a TS managing millions of blocks may
consume enough memory to push it over the limit, but the mm can't do
anything about it (i.e. anchored RAM is 0 across all ops).

JD proposed an alternative: instead of using the root tracker's consumption,
sum up the tablet memtrackers' consumption. That would ensure that the
consumption only reflects anchored RAM (i.e just MRS and DRS), but it's
brittle: there's no guarantee that non-anchored RAM doesn't slip into the
tablet memtracker hierarchy in the future, and searching the tracker tree
for just the tablet trackers is somewhat ugly.

Note that the root tracker will now have a default limit (previously it was
always -1). For the time being this doesn't matter because we're only using
Consume(), never TryConsume(). But in the future we'll transition to
TryConsume() in order to respect the limit, and then it'll matter.

Change-Id: I6a5a23a0120dc6e1a87989fce846b375baf1df27
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6784
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: Adar Dembo <adar@cloudera.com>