Set leader election RPC timeout based on election timeout
When we start a new leader election, we snooze our failure detector for
an election timeout. We should use that same snooze time to set the RPC
timeout for our vote requests. Any votes that come back after the snooze
time will be discounted anyway, since we'll have started a new election.

This prevents election RPCs from piling up in the case that the election
timeout is very short (as in RaftConsensusITest.TestChurnyElections)

I ran TestChurnyElections a few hundred times after this fix in a TSAN
build and it passed.

Change-Id: Ie20ec5463b5bbfa3b51957f6c43b96b2f6779620
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6794
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>