pb_util: WritePBContainerToPath should check for overwriting
Previously only one caller (PathInstanceMetadataFile::Create) needed this,
but I'm going to need it in a second place, so let's build the functionality
into the implementation.

Note that the check is vulnerable to TOCTOU races, but we expect the
permissions on Kudu directories (and all of their path components) to be
sufficiently strict that the races can't be exploited.

Change-Id: Ib09c5ecba0fc0d614d4c5b43ff85febb4efcc8ab
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6806
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>