fs: CreateInitialFileSystemLayout should clean up after itself on failure
We noticed that when the log block manager's hole punching test fails,
subsequent attempts to bring up a server fail with "directory is not empty"
errors, which isn't very useful. Here's an attempt to make this better.

The existing rule is still in effect: filesystem roots must be empty or
non-existent. To make startup idempotent, we'll delete everything we create
(files or directories) on startup.

I tested this by starting a TS on ecryptfs. Predictably, the hole punch test
failed. When I restarted the TS, the test failed again (previously I'd get
an error about the fs root being non-empty).

Change-Id: I9d94005c063905312395b8a38cb5a7cc1192dc40
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6808
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>