log: handle truncated log entry headers
This fixes a rare crash seen in RaftConsensusITest.InsertWithCrashyNodes
in which we'd fail to start up after a log entry header was truncated.

We were handling the case where the log entry itself was truncated, but
if the header was truncated in the middle, we'd return an IOError instead
of a Corruption and thus not continue log recovery.

Change-Id: Ia0509561fdb0e4f4c28e007be6f1c1aea76e4c5e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6811
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>