dist_test: work with latest 'isolate'
The handling of paths in 'isolate' changed upstream so that all
paths must be expressed relative to the isolate file itself. We
were previously putting the isolate files in /tmp, and trying to
use relative paths from the repo directory. This no longer works
in a reasonable way.

Instead, we put the isolate files in build/isolate/ and then we
construct paths relative to this directory.

Tested against ca2862b9255f9cf4e293bd399145a818ce8f04b7 from the
luci-go repository.

Change-Id: I74a4fdf10650fd3dd0752ff321d7923d9accc454
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6815
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>