Unify all the timeouts
The Java client currently has all sorts of timeouts and it's not really clear when one needs to
set them.

This patch reworks it all under the C++ client operation/admin operation timeouts concepts. A
builder pattern is now used to provide those timeouts, which are piped down accordingly into
sessions and scanners.

Change-Id: I3095446fb65b4377ba7f5989d0137ac5a8b5c3bf
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6827
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins