[java client] Add a socket read timeout
Adds a ReadTimeoutHandler to the pipeline, disabled if the passed timeout is 0.
Also removes some isMasterTable cruft.

Tested locally with YCSB and on bolt-80 with ITBLL.

The main deficiency with this patch is that we don't do single RPC timeouts, it just gets
triggered if we haven't read from the socket in so long. That means that an RPC that was sent
10 seconds ago won't get an RPC timeout if other RPCs are properly served on the same Channel.
This seems like a corner case, and the operation timeout itself will trigger anyways. We do set
the timeout in the RPC header correctly.

A structural change in TabletClient is that we now accept that RPCs can be pulled from
rpcs_inflight while decode() is being called. So, if an RPC timeout is triggered on a Channel
so that we start failOrRetry'ing all the in flight RPCs _and_ that concurrently one of them
makes it back, the latter will likely throw a NonRecoverableException. Nothing exceptional
will take place, we're already closing down the TabletClient.

Change-Id: I4701c948b06dd7888ac1c527f5997b1771a669e5
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6828
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>