Stop using incorrect terminology for consensus
This applies a bunch of renames to try to make our consensus terminology more
correct. In particular, we used the term 'quorum' throughout when we really
meant the set of all raft peers, or a particular raft configuration.

This patch was created automatically using a perl script with the exception of
fixing lint errors. The script can be found at:
https://gist.github.com/f6cc588d021c95e3c558

Some of the rewrites aren't perfect -- we should go back through manually to
clean up comments, indentation, etc, where appropriate after this is committed.

Change-Id: I44c581f6ab22ff7672c054da5c5f5d0e8b4c88cb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6822
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>