client: replace remaining gutil usage in public headers with stubs
This last patch turned out to be hard to decompose.

Taking more inspiration from protobuf, let's replace the remaining gutil
usage in the public client headers with a single "stubs" header. To prevent
collisions with our own use of gutil, we'll only do this when building
third party applications (i.e. when NO_STUBS is not defined).

The stubs header provides any macros referenced in the headers, provided
they're not already defined. In doing so, it effectively reverts portions of
commits 2f9281d and 872aba4.

Change-Id: I91a523a85b158855ef97773ad8e42e8e109a9fe3
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6841
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>