tpch: add range-scan support to DAO, refactor scanner
This refactors the TPCH scanner to be a separate class from the DAO.
This allows us to have multiple scanners outstanding at a time (eg
for parallel scan support).

Additionally, I added a new API to open a scanner which performs a
range scan of a subset of the lineitems, based on their primary key.

Change-Id: I4db83ed5603356104b189ce5b3a5b4e9fcd6a515
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6852
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>