Fix EncodedRowKey::IncrementEncodedKey to also increment "raw" keys
Previously, this function had a bug that it would only increment the "encoded"
key, and not the "raw keys". The "raw keys", however, are used for pushdown
into cfile range scans when the key is non-composite.

Additionally, it was using faststring::AdvanceToSuccessor, which is incorrect
in the case of string keys. This function results in 'foo' being advanced to
'fop', but that skips over keys such as 'foobar'.

This patch implements correct logic for incrementing a composite key in a new
row_key-util module. The EncodedKey function now uses these functions to
correctly increment and re-encode a key.

The fault tolerant scan test in client-test didn't catch this bug because it
only scanned against MRS. This patch changes that test to trigger a flush and
verify that fault tolerant scans still behave correctly.

The other fault tolerant scan test in tablet_server-test didn't catch the bug
because it didn't set FLAGS_tablet_server_scan_batch_size_rows to a lower
value. So, even though we asked for small batches, we ended up reading the
entire tablet in a single read in many cases.

Note: EncodedKeyBuilder::BuildSuccessorKey also has a similar bug, which
affects RangePredicateEncoder. That will be fixed separately.

Change-Id: I5a564bbb1ee7e3fafb16cceaf736673d4762825d
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6850
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>